Move start into lib/server
