+++ lib/server/__tests__/start.test.js
@@ -7,9 +7,9 @@
 
 const program = require('commander');
 const openBrowser = require('react-dev-utils/openBrowser');
-const findUnusedPort = require('../findUnusedPort.js');
-const liveReloadServer = require('../server/liveReloadServer.js');
-const server = require('../server/server.js');
+const findUnusedPort = require('../../findUnusedPort.js');
+const liveReloadServer = require('../liveReloadServer.js');
+const server = require('../server.js');
 
 const siteConfig = require(`${process.cwd()}/siteConfig.js`);
 
@@ -20,9 +20,9 @@ jest.mock(`${process.cwd()}/siteConfig.js`, () => jest.fn(), {virtual: true});
 
 jest.mock('commander');
 jest.mock('react-dev-utils/openBrowser');
-jest.mock('../findUnusedPort.js');
-jest.mock('../server/liveReloadServer.js');
-jest.mock('../server/server.js');
+jest.mock('../../findUnusedPort.js');
+jest.mock('../liveReloadServer.js');
+jest.mock('../server.js');
 jest.mock('process');
 
 console.log = jest.fn();
