+++ lib/server/start.js
@@ -7,9 +7,9 @@
 
 const program = require('commander');
 const openBrowser = require('react-dev-utils/openBrowser');
-const findUnusedPort = require('./findUnusedPort.js');
-const liveReloadServer = require('./server/liveReloadServer.js');
-const server = require('./server/server.js');
+const findUnusedPort = require('../findUnusedPort.js');
+const liveReloadServer = require('./liveReloadServer.js');
+const server = require('./server.js');
 
 const CWD = process.cwd();
