+++ lib/start-server.js
@@ -25,7 +25,7 @@ const program = require('commander');
 const CWD = process.cwd();
 const env = require('./server/env.js');
 
-const start = require('./start.js');
+const start = require('./server/start.js');
 
 if (!fs.existsSync(`${CWD}/siteConfig.js`)) {
   console.error(
